Add StatsMapper for mapping statistics data and update DemProcessingService to utilize new mappings

+++ src/main/java/demoanalyzer/com/dem/core/domain/model/stats/StatsKast.java
@@ -1,4 +1,4 @@
 package demoanalyzer.com.dem.core.domain.model.stats;
 
 public record StatsKast(
-    String name, Long steamId, Side side, Integer nRounds, Integer kastRounds, Integer kast) {}
+    String name, Long steamId, Side side, Integer nRounds, Integer kastRounds, Double kast) {}
